React components moved to a separate components directory
